+++ test/upgrade/hosted/unlock_single_package_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2012, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
@@ -8,9 +8,7 @@ import '../../descriptor.dart' as d;
 import '../../test_pub.dart';
 
 void main() {
-  test(
-      "upgrades a locked package's dependers in order to get it to max "
-      'version', () async {
+  test('can unlock a single package only in upgrade', () async {
     await servePackages((builder) {
       builder.serve('foo', '1.0.0', deps: {'bar': '<2.0.0'});
       builder.serve('bar', '1.0.0');
@@ -27,8 +25,21 @@ void main() {
       builder.serve('bar', '2.0.0');
     });
 
+    // This can't upgrade 'bar'
     await pubUpgrade(args: ['bar']);
 
+    await d.appPackagesFile({'foo': '1.0.0', 'bar': '1.0.0'}).validate();
+
+    // Introducing foo and bar 1.1.0, to show that only 'bar' will be upgraded
+    globalPackageServer.add((builder) {
+      builder.serve('foo', '1.1.0', deps: {'bar': '<2.0.0'});
+      builder.serve('bar', '1.1.0');
+    });
+
+    await pubUpgrade(args: ['bar']);
+    await d.appPackagesFile({'foo': '1.0.0', 'bar': '1.1.0'}).validate();
+
+    await pubUpgrade();
     await d.appPackagesFile({'foo': '2.0.0', 'bar': '2.0.0'}).validate();
   });
 }
